refactor code to submit only necessary body for request when editing and fix issue with ordering of tech stack items in edit finalize page
