+++ src/app/(main)/my-voyage/[teamId]/tech-stack/finalize/utils/createFinalList.ts
@@ -1,17 +1,8 @@
-import { getDifferentObject } from "./getDifferentObject";
 import type { SelectedItems } from "@/app/(main)/my-voyage/[teamId]/tech-stack/finalize/types";
 
-function createFinalList(
-  previousSelected: SelectedItems,
-  selectedItems: SelectedItems,
-) {
+function createFinalList(selectedItems: SelectedItems) {
   const finalList = [];
   const keys = Object.keys(selectedItems);
-
-  const differences = getDifferentObject({
-    prev: previousSelected as { [key: number]: number },
-    curr: selectedItems as { [key: number]: number },
-  });
 
   for (let i = 0; i < keys.length; i++) {
     const key = Number(keys[i]);
@@ -26,18 +17,6 @@ function createFinalList(
         ],
       });
     }
-  }
-
-  for (const diff of differences) {
-    finalList.push({
-      categoryId: diff.key,
-      techs: [
-        {
-          techId: diff.previousValue,
-          isSelected: false,
-        },
-      ],
-    });
   }
 
   return finalList;
